fix(ax-fs-ng): complete block lifecycle rollback coverage
