+++ fs/ax-fs-ng/src/block/runtime/lifecycle/tests/batching.rs
@@ -1,0 +1,103 @@
+use super::*;
+
+#[test]
+fn read_blocks_queues_the_next_bounded_window_before_waiting() {
+    let _registrar_guard = lock_test_irq_registrar();
+    crate::os::task::install_test_runtime_ops();
+    install_dma_op(&TEST_DMA_OP);
+    let log = Arc::new(StdMutex::new(Vec::new()));
+    *TEST_IRQ_REGISTRAR.log.lock().unwrap() = Some(log);
+    *TEST_IRQ_REGISTRAR.action.lock().unwrap() = None;
+    TEST_IRQ_REGISTRAR
+        .fail_registration
+        .store(false, Ordering::Release);
+    set_irq_registrar(&TEST_IRQ_REGISTRAR);
+
+    let counters = Arc::new(BatchingQueueCounters::default());
+    let controller = BatchingReadController {
+        queue: Some(BatchingReadQueue {
+            counters: Arc::clone(&counters),
+            next_id: 0,
+            pending: Vec::new(),
+        }),
+    };
+    let irq = IrqId::new(IrqDomainId(1), HwIrq(12));
+    let handle = BlockDeviceHandle::start(RdifBlockDevice::new_with_irqs(
+        "batching-read",
+        [BlockIrqSource { source_id: 0, irq }],
+        Box::new(controller),
+    ))
+    .unwrap();
+
+    let reader = Arc::clone(&handle);
+    let (result_tx, result_rx) = mpsc::channel();
+    let read_thread = thread::spawn(move || {
+        let mut buffer = vec![0; 8 * 512];
+        result_tx.send(reader.read_blocks(0, &mut buffer)).unwrap();
+    });
+    let deadline = Instant::now() + Duration::from_secs(1);
+    while counters.submitted.load(Ordering::Acquire) < 4 {
+        assert!(
+            Instant::now() < deadline,
+            "synchronous wrapper waited before submitting the full I/O window"
+        );
+        thread::yield_now();
+    }
+    while counters.commits.load(Ordering::Acquire) < 1 {
+        assert!(
+            Instant::now() < deadline,
+            "maintenance task did not commit the submitted I/O window"
+        );
+        thread::yield_now();
+    }
+
+    assert_eq!(counters.largest_batch.load(Ordering::Acquire), 4);
+    assert_eq!(counters.commits.load(Ordering::Acquire), 1);
+    while handle
+        .inner
+        .cpu_channels
+        .lock()
+        .iter()
+        .map(|channel| channel.channel.queued_len())
+        .sum::<usize>()
+        < 4
+    {
+        assert!(
+            Instant::now() < deadline,
+            "the requester did not queue the second window before the first IRQ"
+        );
+        thread::yield_now();
+    }
+    assert_eq!(counters.submitted.load(Ordering::Acquire), 4);
+    assert!(result_rx.try_recv().is_err());
+    assert_eq!(
+        TEST_IRQ_REGISTRAR.run_registered_action(),
+        BlockIrqOutcome::Wake
+    );
+    while counters.submitted.load(Ordering::Acquire) < 8 {
+        assert!(
+            Instant::now() < deadline,
+            "maintenance task did not refill from the queued second window"
+        );
+        thread::yield_now();
+    }
+    while counters.commits.load(Ordering::Acquire) < 2 {
+        assert!(
+            Instant::now() < deadline,
+            "maintenance task did not commit the refilled second window"
+        );
+        thread::yield_now();
+    }
+    assert_eq!(
+        TEST_IRQ_REGISTRAR.run_registered_action(),
+        BlockIrqOutcome::Wake
+    );
+    assert!(
+        result_rx
+            .recv_timeout(Duration::from_secs(1))
+            .unwrap()
+            .is_ok()
+    );
+    read_thread.join().unwrap();
+    assert_eq!(handle.shutdown(), 1);
+}
