+++ scripts/axbuild/src/test/std.rs
@@ -25,6 +25,7 @@ pub(crate) struct StdTestArgs {
 #[derive(Clone, Copy, Debug)]
 struct PackageFeatureProfile {
     name: &'static str,
+    no_default_features: bool,
     features: &'static [&'static str],
     name_filter: Option<&'static str>,
     expected_tests: &'static [&'static str],
@@ -33,6 +34,7 @@ struct PackageFeatureProfile {
 const AX_TASK_FEATURE_PROFILES: &[PackageFeatureProfile] = &[
     PackageFeatureProfile {
         name: "host-test+multitask+irq-pure",
+        no_default_features: false,
         features: &["host-test", "multitask", "irq"],
         name_filter: Some("std_tests::"),
         expected_tests: &[
@@ -44,6 +46,7 @@ const AX_TASK_FEATURE_PROFILES: &[PackageFeatureProfile] = &[
     },
     PackageFeatureProfile {
         name: "host-test+multitask-task-initialization",
+        no_default_features: false,
         features: &["host-test", "multitask"],
         name_filter: Some(TASK_INITIALIZATION_FILTER),
         expected_tests: &["api::tests::task_initialization_precedes_scheduling"],
@@ -52,6 +55,7 @@ const AX_TASK_FEATURE_PROFILES: &[PackageFeatureProfile] = &[
 
 const AX_DRIVER_FEATURE_PROFILES: &[PackageFeatureProfile] = &[PackageFeatureProfile {
     name: "host-test+rtc+starfive-jh7110-dwmmc",
+    no_default_features: false,
     features: &["host-test", "rtc", "starfive-jh7110-dwmmc"],
     name_filter: None,
     expected_tests: &[],
@@ -59,6 +63,7 @@ const AX_DRIVER_FEATURE_PROFILES: &[PackageFeatureProfile] = &[PackageFeaturePro
 
 const HOST_TEST_FEATURE_PROFILES: &[PackageFeatureProfile] = &[PackageFeatureProfile {
     name: "host-test",
+    no_default_features: false,
     features: &["host-test"],
     name_filter: None,
     expected_tests: &[],
@@ -66,6 +71,7 @@ const HOST_TEST_FEATURE_PROFILES: &[PackageFeatureProfile] = &[PackageFeaturePro
 
 const ALLOC_FEATURE_PROFILES: &[PackageFeatureProfile] = &[PackageFeatureProfile {
     name: "alloc",
+    no_default_features: false,
     features: &["alloc"],
     name_filter: None,
     expected_tests: &[],
@@ -73,6 +79,7 @@ const ALLOC_FEATURE_PROFILES: &[PackageFeatureProfile] = &[PackageFeatureProfile
 
 const FS_FEATURE_PROFILES: &[PackageFeatureProfile] = &[PackageFeatureProfile {
     name: "fs",
+    no_default_features: false,
     features: &["fs"],
     name_filter: None,
     expected_tests: &[],
@@ -80,10 +87,98 @@ const FS_FEATURE_PROFILES: &[PackageFeatureProfile] = &[PackageFeatureProfile {
 
 const STARRY_KERNEL_FEATURE_PROFILES: &[PackageFeatureProfile] = &[PackageFeatureProfile {
     name: "std-tests-only",
+    no_default_features: false,
     features: &[],
     name_filter: Some("std_tests::"),
     expected_tests: &[],
 }];
+
+const AX_FS_NG_FEATURE_PROFILES: &[PackageFeatureProfile] = &[
+    PackageFeatureProfile {
+        name: "host-test",
+        no_default_features: false,
+        features: &["host-test"],
+        name_filter: None,
+        expected_tests: &[],
+    },
+    PackageFeatureProfile {
+        name: "host-test-resource-rollback-discovery",
+        no_default_features: false,
+        features: &["host-test"],
+        name_filter: Some("until_controller_shutdown"),
+        expected_tests: &[
+            "block::runtime::lifecycle::tests::resource_rollback::duplicate_queue_update_keeps_current_and_trailing_queues_until_controller_shutdown",
+            "block::runtime::lifecycle::tests::resource_rollback::failed_hctx_start_keeps_current_and_trailing_queues_until_controller_shutdown",
+            "block::runtime::lifecycle::tests::resource_rollback::rejected_device_info_update_keeps_emitted_queue_until_controller_shutdown",
+        ],
+    },
+    PackageFeatureProfile {
+        name: "host-test-ready-publication-discovery",
+        no_default_features: false,
+        features: &["host-test"],
+        name_filter: Some("ready_device_rejects_changed_device_info_without_overwriting_epoch"),
+        expected_tests: &[
+            "block::runtime::lifecycle::tests::publication::ready_device_rejects_changed_device_info_without_overwriting_epoch",
+        ],
+    },
+];
+
+const NVME_FEATURE_PROFILES: &[PackageFeatureProfile] = &[
+    PackageFeatureProfile {
+        name: "default",
+        no_default_features: false,
+        features: &[],
+        name_filter: None,
+        expected_tests: &[],
+    },
+    PackageFeatureProfile {
+        name: "rearm-state-discovery",
+        no_default_features: false,
+        features: &[],
+        name_filter: Some("rearm_during_initialization_preserves_waiting_for_irq_state"),
+        expected_tests: &[
+            "block::tests::rearm_during_initialization_preserves_waiting_for_irq_state",
+        ],
+    },
+];
+
+const SDMMC_RDIF_FEATURE_PROFILES: &[PackageFeatureProfile] = &[
+    PackageFeatureProfile {
+        name: "rdif",
+        no_default_features: true,
+        features: &["rdif"],
+        name_filter: None,
+        expected_tests: &[],
+    },
+    PackageFeatureProfile {
+        name: "rdif-lifecycle-discovery",
+        no_default_features: true,
+        features: &["rdif"],
+        name_filter: Some("ready_online_smp_repeats_info_without_reissuing_resources"),
+        expected_tests: &[
+            "sdio::tests::rdif_lifecycle::ready_online_smp_repeats_info_without_reissuing_resources",
+        ],
+    },
+];
+
+const AXBUILD_FEATURE_PROFILES: &[PackageFeatureProfile] = &[
+    PackageFeatureProfile {
+        name: "default",
+        no_default_features: false,
+        features: &[],
+        name_filter: None,
+        expected_tests: &[],
+    },
+    PackageFeatureProfile {
+        name: "ax-fs-ng-axtest-feature-discovery",
+        no_default_features: false,
+        features: &[],
+        name_filter: Some("axfs_ng_axtest_excludes_the_host_sync_backend"),
+        expected_tests: &[
+            "build::tests::std_features::axfs_ng_axtest_excludes_the_host_sync_backend",
+        ],
+    },
+];
 
 #[derive(Clone, Copy, Debug, Eq, Hash, PartialEq)]
 enum CargoTestAction {
@@ -94,6 +189,7 @@ enum CargoTestAction {
 #[derive(Clone, Debug, Eq, Hash, PartialEq)]
 struct CargoTestInvocation {
     package: String,
+    no_default_features: bool,
     features: Vec<String>,
     name_filter: Option<String>,
     action: CargoTestAction,
@@ -103,6 +199,7 @@ impl CargoTestInvocation {
     fn default_for(package: &str) -> Self {
         Self {
             package: package.to_owned(),
+            no_default_features: false,
             features: Vec::new(),
             name_filter: None,
             action: CargoTestAction::Run,
@@ -116,6 +213,7 @@ impl CargoTestInvocation {
     ) -> Self {
         Self {
             package: package.to_owned(),
+            no_default_features: profile.no_default_features,
             features: profile
                 .features
                 .iter()
@@ -128,6 +226,9 @@ impl CargoTestInvocation {
 
     fn args(&self) -> Vec<String> {
         let mut args = vec!["test".into(), "-p".into(), self.package.clone()];
+        if self.no_default_features {
+            args.push("--no-default-features".into());
+        }
         if !self.features.is_empty() {
             args.push("--features".into());
             args.push(self.features.join(","));
@@ -351,9 +452,13 @@ fn package_feature_profiles(package: &str) -> Option<&'static [PackageFeaturePro
         | "ax-net"
         | "dma-api"
         | "buddy-slab-allocator" => Some(HOST_TEST_FEATURE_PROFILES),
+        "ax-fs-ng" => Some(AX_FS_NG_FEATURE_PROFILES),
         "ax-io" | "axbacktrace" => Some(ALLOC_FEATURE_PROFILES),
         "ax-task" => Some(AX_TASK_FEATURE_PROFILES),
         "ax-driver" => Some(AX_DRIVER_FEATURE_PROFILES),
+        "nvme-driver" => Some(NVME_FEATURE_PROFILES),
+        "sdmmc-protocol" => Some(SDMMC_RDIF_FEATURE_PROFILES),
+        "axbuild" => Some(AXBUILD_FEATURE_PROFILES),
         "axvisor" => Some(FS_FEATURE_PROFILES),
         "starry-kernel" => Some(STARRY_KERNEL_FEATURE_PROFILES),
         _ => None,
@@ -523,9 +628,14 @@ mod tests {
             self
         }
 
-        fn with_listing(mut self, profile: &PackageFeatureProfile, tests: &[&str]) -> Self {
+        fn with_listing(
+            mut self,
+            package: &str,
+            profile: &PackageFeatureProfile,
+            tests: &[&str],
+        ) -> Self {
             self.results.insert(
-                CargoTestInvocation::for_profile("ax-task", profile, CargoTestAction::List),
+                CargoTestInvocation::for_profile(package, profile, CargoTestAction::List),
                 CargoRunOutput {
                     success: true,
                     stdout: render_test_list(tests),
@@ -536,7 +646,20 @@ mod tests {
 
         fn with_ax_task_discovery(mut self) -> Self {
             for profile in AX_TASK_FEATURE_PROFILES {
-                self = self.with_listing(profile, profile.expected_tests);
+                self = self.with_listing("ax-task", profile, profile.expected_tests);
+            }
+            self
+        }
+
+        fn with_profile_discovery(
+            mut self,
+            package: &str,
+            profiles: &[PackageFeatureProfile],
+        ) -> Self {
+            for profile in profiles {
+                if !profile.expected_tests.is_empty() {
+                    self = self.with_listing(package, profile, profile.expected_tests);
+                }
             }
             self
         }
@@ -740,6 +863,115 @@ mod tests {
     }
 
     #[test]
+    fn lifecycle_packages_select_full_and_discovery_profiles() {
+        let root = PathBuf::from("/tmp/workspace");
+        let packages = [
+            "ax-fs-ng",
+            "ahci-driver",
+            "nvme-driver",
+            "rdif-block",
+            "sdmmc-protocol",
+            "axbuild",
+        ]
+        .map(str::to_string)
+        .to_vec();
+        let mut runner = FakeCargoRunner::succeeding()
+            .with_profile_discovery("ax-fs-ng", AX_FS_NG_FEATURE_PROFILES)
+            .with_profile_discovery("nvme-driver", NVME_FEATURE_PROFILES)
+            .with_profile_discovery("sdmmc-protocol", SDMMC_RDIF_FEATURE_PROFILES)
+            .with_profile_discovery("axbuild", AXBUILD_FEATURE_PROFILES);
+
+        let failed = run_std_tests(&mut runner, &root, &packages).unwrap();
+
+        assert!(failed.is_empty());
+        let invocations = runner
+            .invocations
+            .iter()
+            .map(|(_, invocation)| invocation)
+            .collect::<Vec<_>>();
+        assert!(invocations.contains(&&CargoTestInvocation::for_profile(
+            "ax-fs-ng",
+            &AX_FS_NG_FEATURE_PROFILES[0],
+            CargoTestAction::Run,
+        )));
+        assert!(invocations.contains(&&CargoTestInvocation::default_for("ahci-driver")));
+        assert!(invocations.contains(&&CargoTestInvocation::for_profile(
+            "nvme-driver",
+            &NVME_FEATURE_PROFILES[0],
+            CargoTestAction::Run,
+        )));
+        assert!(invocations.contains(&&CargoTestInvocation::default_for("rdif-block")));
+        assert!(invocations.contains(&&CargoTestInvocation::for_profile(
+            "sdmmc-protocol",
+            &SDMMC_RDIF_FEATURE_PROFILES[0],
+            CargoTestAction::Run,
+        )));
+        assert!(invocations.contains(&&CargoTestInvocation::for_profile(
+            "axbuild",
+            &AXBUILD_FEATURE_PROFILES[0],
+            CargoTestAction::Run,
+        )));
+    }
+
+    #[test]
+    fn sdmmc_rdif_profiles_disable_default_features() {
+        for profile in SDMMC_RDIF_FEATURE_PROFILES {
+            let invocation =
+                CargoTestInvocation::for_profile("sdmmc-protocol", profile, CargoTestAction::Run);
+            let args = invocation.args();
+
+            assert_eq!(
+                &args[..6],
+                [
+                    "test",
+                    "-p",
+                    "sdmmc-protocol",
+                    "--no-default-features",
+                    "--features",
+                    "rdif",
+                ]
+            );
+        }
+    }
+
+    #[test]
+    fn lifecycle_discovery_profiles_name_every_critical_regression() {
+        let expected = AX_FS_NG_FEATURE_PROFILES
+            .iter()
+            .chain(NVME_FEATURE_PROFILES)
+            .chain(SDMMC_RDIF_FEATURE_PROFILES)
+            .chain(AXBUILD_FEATURE_PROFILES)
+            .flat_map(|profile| profile.expected_tests.iter().copied())
+            .collect::<BTreeSet<_>>();
+
+        assert!(expected.contains(
+            "block::runtime::lifecycle::tests::resource_rollback::duplicate_queue_update_keeps_current_and_trailing_queues_until_controller_shutdown"
+        ));
+        assert!(expected.contains(
+            "block::runtime::lifecycle::tests::resource_rollback::failed_hctx_start_keeps_current_and_trailing_queues_until_controller_shutdown"
+        ));
+        assert!(expected.contains(
+            "block::runtime::lifecycle::tests::resource_rollback::rejected_device_info_update_keeps_emitted_queue_until_controller_shutdown"
+        ));
+        assert!(expected.contains(
+            "block::runtime::lifecycle::tests::publication::ready_device_rejects_changed_device_info_without_overwriting_epoch"
+        ));
+        assert!(
+            expected.contains(
+                "block::tests::rearm_during_initialization_preserves_waiting_for_irq_state"
+            )
+        );
+        assert!(expected.contains(
+            "sdio::tests::rdif_lifecycle::ready_online_smp_repeats_info_without_reissuing_resources"
+        ));
+        assert!(
+            expected.contains(
+                "build::tests::std_features::axfs_ng_axtest_excludes_the_host_sync_backend"
+            )
+        );
+    }
+
+    #[test]
     fn ax_task_uses_pure_and_task_initialization_feature_profiles() {
         let root = PathBuf::from("/tmp/workspace");
         let packages = vec!["ax-task".to_string()];
@@ -913,8 +1145,9 @@ mod tests {
         let initialization_profile = &AX_TASK_FEATURE_PROFILES[1];
         let mut runner = FakeCargoRunner::succeeding()
             .with_ax_task_discovery()
-            .with_listing(pure_profile, &["api::std_tests::unexpected"])
+            .with_listing("ax-task", pure_profile, &["api::std_tests::unexpected"])
             .with_listing(
+                "ax-task",
                 initialization_profile,
                 initialization_profile.expected_tests,
             );
